Archive funtion added, Dashboard Improved
destroyed rentals are now archived into a csv filed and then destroyed. Changed buttons to more accurately describe this behavior. Dashboard now works as a hub for all the other pages and gives brief tutorials on how to work each page. The system is pretty intuitive, but might expand on these helpers.